+++ HISTORY.rst
@@ -10,6 +10,7 @@ Release History
 * GridPoint now Hashable
 * Summit and Saddle Objects now have Neighbor Finding functions
 * improvements to rebuildSaddles such as fast bailouts and fixed several bugs where calculated midpoints were being placed outside of multipoints. This usually happens when two linked high shores are on the same plane.
+* base_coordinate and child objects now include dms displays.
 
 0.5.10 (September 17, 2018)
 +++++++++++++++++++++++++++